Move the server class inside lib folder
